convert to proper custom resources
removes crazyballs inheritance that was only causing issues.
