--- libraries/etcd_base.rb
@@ -1,16 +1,0 @@
-module EtcdCookbook
-  class EtcdBase < Chef::Resource
-    require 'helpers_base'
-
-    ################
-    # Type Constants
-    ################
-
-    Boolean = property_type(
-      is: [true, false],
-      default: false
-    ) unless defined?(Boolean)
-
-    declare_action_class
-  end
-end
